Views and autos now added/removed in Txn
Before, we had separate locking for the revision/transaction system and
adding/removing views. We now handle everything in transactions -
views/autos are added and removed during the transaction, and if the
transaction succeeds and commits then this is reflected in the shelf.
Note that the list of views/autos is not actually in the Revision, but
only because it's not actually useful for anything to read apart from
Shelf - could be moved to Revision later without much difficulty if
there was a use case for this.
Note that this means that views/autos are first applied on the revision
created by the transaction where they are added - that is, we now know
exactly which revision they will see first. When nested transactions
are added, we will also be able to handle the currently problematic
case of adding views in a transaction, without risking having the views
run on an older transaction than the one they are created in. For now,
we just have the normal nested transaction problems rather than the old
weird extra problem.