Add nested Txns, fix reaction target tracking, add tests
TransactDefault implementation now tracks the Txn being used in the
current thread, and executes any nested transactions in it. Previously,
nested transactions would create a new transaction. If they then tried
to access any new Boxes in scope from the lexically enclosing
transaction they would error, since the nested transaction was looking
at the current revision where those boxes don't yet exist. Now, if a
thread already has a transaction, any new transact function is run in
the existing transaction (merged). This makes it much easier to track
transactions, at the cost of requiring that use of a transaction is
single threaded (which is the normal use case).

Reaction target tracking was previously incorrect when a reaction sets
a value into a box which is the same as its old value. Before, this
skipped writing entirely, so didn't call Reactor.afterSet(...). Now,
afterSet is always called, but a boolean is passed to indicate if an
actual change was made. This allows targets of reactions always to be
registered, but we don't bother re-evaluating reactions having the Box
as a source when the Box value doesn't actually change.

Finally, added tests from boxes-sbt and started porting old
non-transactional tests from boxes-core.