Better Delta
D now represents a single delta which could be composed of multiple
constituent deltas, using Monoid[D]. This means we don't need to use
flatMap or require a List, and could make D more efficient by allowing
it to combine constituent deltas when instances are  appended with
Monoid.

Updated example to use this system with a MapDeltas example for D. This
is essentially the system now used for revisions and their deltas,
however to make things less confusing and possibly more efficient we
use the specific types in revision - this might change later.