Better monadic Revision system
We now use a system like the example Delta monad to require that
functions return a delta to state rather than a new state, making
things safer and also allowing much neater access to the ordered list
of basic operations on the revision, which should be useful for
reactions. This works nearly identically from the user's perspective.