Implement reactions in free
Ported nearly all of the old Reactor system to free implementation.
This is much neater due to lack of "callbacks", however it's still
imperative for the actual application of reactions, this should be
fixed in future but for now shouldn't be visible, no mutable state is
shared, no threading used etc.

Only "changedSources" is left to reimplement, this should fit nicely
into the interpreter.