Add "just" to BoxDeltaF
In the API we generally expect a BoxScript instead of plain values,
since any access to actual Box contents etc. does require a BoxScript
to work. JustF allows for use of a plain value to be passed as a
BoxScript. This probably isn't often much use at all, but is easy
enough to provide.