Fix issue with multiple BoxChanges for same Box in same Revision
Moving to a list of deltas and applying them all means that when we
make a new Revision, we may make multiple BoxChange instances for a
single box in the new revision. Previously, these were case classes,
and so all changes with the same revision index were equal. This meant
that in Box, a new change might not replace an old equal one in the
weak hash map of changes to values. This resulted in both the
possibility of an incorrect value, and the possibility that a
Revision's map would refer to a BoxChange not actually stored in the
Box. This has been fixed by using a normal class for BoxChange, so all
instances are unequal. Since BoxChange is not in public API, this
doesn't change any behaviour, just fixes the issue.