Rename BoxDelta cases, better contents
Rename cases to past tense to avoid confusion with operations. These
now contain all relevant data - WriteBox is now BoxWritten and has both
old and new values, and BoxCreated also stores the initial value.
BoxRead stores the value that was actually read. This makes it easier
to see what is happening in a list of deltas.