Remove cal, BoxM based replacement for PathReaction
BoxScriptImports.cal can be entirely replaced using BoxR, removing the
need for a Box that exists just to be updated.
Working on a replacement for PathReaction that uses BoxM, and again
removes a Box that exists just to contain the calculated path value -
will remove PathReaction when this is tested.