Make Observer more flexible
Can now optionally specify a different executor to be used for the
effect of a script. This defaults to None and so runs the script and
effect in the same executor (and same runnable on that executor),
giving the old behaviour. The new behaviour is useful mainly for Swing
observers, so that the script can be run off the Swing thread, then
invoke later used to perform the effect that actually interacts with
Swing.