Reactor ignores pointless writes for reaction propagation
When a value is written to the same value it already has, this does not
cause an execution of reactions sourcing that value. However when a
reaction writes to a box, even if it writes the same value that box has
already, it is still considered to target that box.

This will increase efficiency by only running reactions when their
sources have actually changed.

This also means that the BoxScriptInterpreter does not need to decide
whether to pass write deltas to the Reactor - it should always do so,
and Reactor can ignore them as appropriate.