Add optional default selection as Option or Set
When the selection would become an empty Set or None, we can instead
create a new selection based on the current List using a provided
function. Default is to leave selection as empty Set or None. Functions
are provided to select the first or all elements as appropriate.
Add tests for default selection.