Added EndToken. Add Replaces typeclass.
It is useful to be able to see the end of the token stream, so add an
EndToken - this must be returned when no more real tokens are available
(as many times as requested). This replaces throwing a
NoTokenException. NoTokenException has been removed. EndTokens do
nothing when written.

Replaces typeclass allows us to traverse an object graph, and when we
find a particular Box, replace its contents with data read from tokens.
This is useful for e.g. network comms, where we can accept a box id and
some json data, and use the data to set a new box value.

Add checks for EndToken in JsonSpec and ProtobufSpec, to check the
corresponding TokenReaders work as expected.