Make JsonTokenReader more flexible, especially for casting
Split up json reader into two classes - JsonDirectTokenReader will just
perform minimal json parsing, and produce the "widest" data types,
essentially BigDecimalToken or BigIntegerToken for numbers, as
appropriate. JsonTokenReader now just wraps another TokenReader, and
applies casting for pullLong, pullFloat etc. to convert the
BigDecimalTokens and BigIntegerTokens to the requested primitive type
when this is safe. The actual casting is split out into utility
functions in JsonCasting.

This allows (for example) wrapping of a BufferTokenReader holding
tokens parsed from a JsonDirectTokenReader, so that the tokens are
still converted appropriately. This is useful for box-http4s, which
needs to pre-process received tokens to strip out "header" information
before passing the tokens on to a Format.

This means that JsonTokenReader.apply() methods should be used instead
of new JsonTokenReader, which now expects a TokenReader not a Reader -
updated JsonSpec for this.