Add Modifies typeclass
This is now part of Format typeclass, and is similar to Replaces, but
more general. It allows for boxes to be modified using tokens read in a
BoxReaderScript. This can be used for example to allow for a Node in a
Box to be modified. The modifications allowed are fairly general - they
receive a Box and must produce a BoxReaderScript, but that's the only
requirement. A simple way of using this is with the new Action[T] trait
- this just provides an act method accepting a box and producing a
script as required. This can be used with ActionFormats to produce a
new format based on a Format[T] and a Format[Action[T]] that will
provide for using that action to modify T.

This is really only useful for a kind of RPC, where we accept tokenised
Actions and a Box id, and we can then apply that action to that box in
a type-safe way, with actions provided by typeclasses making the system
fairly extensible.