New use of Modifies for Node, new defaults for ids
Modifies should be much easier to use if we make it apply to the Node
containing a box itself, rather than using modifyBox to apply the
modification to a contained Node. This means that we can apply
modifications to any Node with a box (a node without a box is just a
case object, which can't be modified). This is slightly messy since it
is using an (arbitrary) box of a Node as an identifier for the Node,
however this has the advantage that it will work with any Node anywhere
in the data graph, as long as that Node follows the requirements we
already have (that it doesn't share any Box with any other Node).

Adapt Formats to work this way, Action now acts on T not Box[T].

Adapt ModifiesSpec, and add a test for a Node in a list.

Make IdLinks the default boxLinkStrategy for NodeFormats, since this is
nearly always right - ids can be easily ignored if not needed.

Fix TaggedUnionFormats - still needs to be tested with a spec.