Removed linkref. AssignId is now GetId.
Working towards a system that only has ids, no references to them. This
implies we can't preserve the identicality of nodes etc. that appear in
different places in the data. However this is normal in e.g. json, and
removes the need for unsafe casts of "cached" instances when using a
ref.